epps_workspace/requirements.txt
pydantic>=2.4.0
pyyaml>=6.0
openai>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
